Reorganize Heatmap into static and interactive versions
